server/requirements.txt
websockets>=11.0.3
google-api-python-client>=2.90.0
pylint-protobuf>=0.20.2